+++ tests/logtalk/directives/include_1/include_3.pl
@@ -1,0 +1,2 @@
+
+:- dynamic(bar/1).